+++ contrib/openai/image.go
@@ -6,7 +6,6 @@ import (
 	"errors"
 	"fmt"
 	"strconv"
-	"strings"
 
 	"github.com/go-kratos/blades"
 	"github.com/openai/openai-go/v2"
@@ -68,34 +67,6 @@ func (p *ImageProvider) NewStream(ctx context.Context, req *blades.ModelRequest,
 		return nil
 	})
 	return pipe, nil
-}
-
-func promptFromMessages(messages []*blades.Message) (string, error) {
-	if len(messages) == 0 {
-		return "", ErrPromptRequired
-	}
-	var sections []string
-	for _, msg := range messages {
-		switch msg.Role {
-		case blades.RoleSystem, blades.RoleUser:
-			var textParts []string
-			for _, part := range msg.Parts {
-				switch v := part.(type) {
-				case blades.TextPart:
-					if strings.TrimSpace(v.Text) != "" {
-						textParts = append(textParts, v.Text)
-					}
-				}
-			}
-			if len(textParts) > 0 {
-				sections = append(sections, strings.Join(textParts, "\n"))
-			}
-		}
-	}
-	if len(sections) == 0 {
-		return "", ErrPromptRequired
-	}
-	return strings.Join(sections, "\n\n"), nil
 }
 
 func applyImageOptions(params *openai.ImageGenerateParams, model string, cfg blades.ImageOptions) {
